generation of a small amount of preliminary Javadoc (more planned)
reimplementation of config-file writing using objects, instead of static classes and static methods, in order to reduce the passing around of a bunch of parameters, and simplify the overall scheme

some renaming:  most importantly:  a Simulation is now the top-level object, and it has a SimulationSettings and a Model.  The Model now has no knowledge of the SimulationSettings, and vice versa:  any information that belongs together according to Smoldyn, but should be split across the Model and SimulationSettings, will be remembered as belonging together by hash tables owned by the Simulation.

git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@6824 d59c7341-666b-7e48-8bb3-2014e8cfb900