put default values for InternalSettings (that were previously only in JavaDoc comments) into the code, and removed one parameter that was essentially a synonym (molperbox)
support for surface actions now somewhat complete, but relatively untested.  also, the design is very poor -- since actions need to be specified for both surfaces and species, it's hard to search through the data structure in an intelligent way.  what if the user wants to change all actions by species?  fine, no problem, but what if the user also needs to change all actions by surface?  then there's a problem that I wasn't able to solve -- I didn't want to simply use an arraylist to hold a bunch of actions, because then there could be duplicates, and then it would be a huge mess and probably hard to maintain or change, so the solution is just to use nested hashtables, so: hash by surface, then by species.  the client can then get all actions for a specific surface, but getting all actions for a species is more complicated

git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@6990 d59c7341-666b-7e48-8bb3-2014e8cfb900