Deploying:
Changes to ROI and ROImanager to make ROIs aware of which image they were drawn on or opened into from a file.  This allows corrections for differences in CZT dimensions when hopping an ROI from the manager into a different hyperstack window in GLOWormJ.  Shoul enhance the ability to highlight a region in space/time in one movie based on annotations made on another movie.

git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@8682 d59c7341-666b-7e48-8bb3-2014e8cfb900