Bug 4010 - EBChombo: write files in parallel
only in parallel, the ghost for membrane index level data is one because we need to exchange. Since in serial there is no need to exchange and we don't exchange, we set the ghost to zero so that when we collect membrane index to create the map, we grow by zero, so that we don't accidently include unnecessary points.

git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@16569 d59c7341-666b-7e48-8bb3-2014e8cfb900